feat: create container component app provider
